+++ eras/alonzo/impl/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### `testlib`
 
+* Export `fixupRedeemerIndices` from `Alonzo.ImpTest`
 * Enhance the `addCollateralInput` fixup function to not add collateral if it's already present
 
 ## 1.10.0.0